--- app/src/main/java/edu/nextstep/camp/calculator/model/const.kt
@@ -1,3 +1,0 @@
-package edu.nextstep.camp.calculator.model
-
-const val EMPTY = ""
