+++ domain/src/main/java/edu/nextstep/camp/calculator/domain/calculator/FormulaElement.kt
@@ -1,4 +1,4 @@
-package edu.nextstep.camp.calculator.domain
+package edu.nextstep.camp.calculator.domain.calculator
 
 sealed class FormulaElement
 sealed class Operator: FormulaElement() {
