+++ domain/src/main/java/edu/nextstep/camp/calculator/domain/calculator/StringCalculator.kt
@@ -1,4 +1,4 @@
-package edu.nextstep.camp.calculator.domain
+package edu.nextstep.camp.calculator.domain.calculator
 
 class StringCalculator {
     companion object {
