+++ domain/src/main/java/edu/nextstep/camp/calculator/domain/const.kt
@@ -1,0 +1,3 @@
+package edu.nextstep.camp.calculator.domain
+
+const val EMPTY = ""
