+++ domain/src/test/kotlin/edu/nextstep/camp/calculator/domain/StringCalculatorTest.kt
@@ -1,6 +1,7 @@
 package edu.nextstep.camp.calculator.domain
 
 import com.google.common.truth.Truth
+import edu.nextstep.camp.calculator.domain.calculator.StringCalculator
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
